+++ src/lint/rules/tests/common.rs
@@ -34,7 +34,7 @@ pub fn run_linter(source_code: &str, rules: &CurrentRules)
     print!("\nSnippet to test on:\n{}\n", source_code);
     let ast = create_ast_from_snippet(source_code);
     print!("Resulting AST:\n{:#?}\n", ast);
-    begin_style_check(ast, source_code.to_string(), rules)
+    begin_style_check(ast, source_code, rules)
 }
 
 pub fn assert_snippet(source_code: &str, expected_errors: usize, rules: &CurrentRules) {
@@ -48,8 +48,8 @@ pub fn robust_assert_snippet(source_code: &str, expected_errors: Vec<ExpectedDML
     let lint_errors = run_linter(source_code, rules).unwrap();
     assert_eq!(lint_errors.len(), expected_errors.len(), "{:#?}", lint_errors);
 
-    let actual_errors: HashSet<_> = lint_errors.iter().map(|e| (e.error.range, e.rule_type.clone())).collect();
-    let expected_errors: HashSet<_> = expected_errors.iter().map(|e| (e.range, e.rule_type.clone())).collect();
+    let actual_errors: HashSet<_> = lint_errors.iter().map(|e| (e.error.range, e.rule_type)).collect();
+    let expected_errors: HashSet<_> = expected_errors.iter().map(|e| (e.range, e.rule_type)).collect();
 
     assert_eq!(actual_errors, expected_errors, "Mismatch between actual and expected errors:\nActual: {:#?}\nExpected: {:#?}", actual_errors, expected_errors);
 }