+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   has no semantic effect.
 - The parameter declaration patterns ":=" and ": \<type\> = \<val\>" are now
   correctly parsed.
+- You can now annotate dml source to disable reporting of specific lints for specific files or lines,
+  see [USAGE.md](USAGE.md) for instructions on how to use it
 
 ## 0.9.12
 - Added 'simics\_util\_vect' as a known provisional (with no DLS semantics)